requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
sqlalchemy==2.0.23
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.28.1
google-generativeai==0.3.2
